feat: Upgrade to next.config.ts
